Bugfix for issues with gush-control
The bug was caused by the fact that "server" command is defined by gush-control
gem which was not required by default by gush (which is reasonable) and was
loaded when other files in the project were required from Gushfile.rb.

Previously Gushfile was loaded from the binary itself, but now it is no longer
possible, as '-f' flag specifies the location of that file.

This commit works around that issue by loading everything specified in the
gemfile, so that gush-control will be loaded when it's available