Removed GPL libraries in favor of MIT compatible ones
Removed the [colorize](https://github.com/fazibear/colorize) and
[ruby-graphviz](https://github.com/glejeune/Ruby-Graphviz) libraries
which were GPL and not compatible with the Gush MIT license. Replaced
them with [paint](https://github.com/janlelis/paint) and [graphviz](https://github.com/ioquatix/graphviz),
respectively.

Enhanced options for graph vizualization by allowing a filename to be
specified to retain the output of the graph. Allowed graphs to be
generated from specific workflows that exist in Redis in addition to
coming straight from the class definition (for systems that may choose
to make the structure of the graph dynamic based on inputs).

Standardized CLI [documentation
format](https://developers.google.com/style/code-syntax).