Add globals to Workflow and Job
For shared functionality across Workflow classes that depends on some variable
data, it's useful to have named parameters for the Workflow that are
automatically forwarded to all Job instances.

This commit adds a `globals` keyword arg to the Workflow initializer, which
should be a hash that is then stored in a `globals` hash attribute, is
persisted, and is merged into the `params` sent to each Job instance.