Added `sync` option to job configuration to run job synchronously / immediately
There might be a case when you want some jobs to complete immediately after a
call to `Workflow#start!`. For example, some jobs might complete very quickly
and it can be helpful to show the results of those jobs in the response to some
web request that started the workflow. Any job with the `sync: true` option is
not sent to the queuing adapter but directly executed by blocking the execution
of others until it’s finished.