Persist dependencies so that `Workflow#configure` is not required on load
Previously, `Workflow#configure` was called every time a Workflow was
instantiated. This could create broken dependency graphs, e.g. if a
configure method sets up job dependencies based on some mutable data like
a timestamp argument or a database value.

Instead, serialize workflow dependencies along with the rest of a workflow's
data and reload it via `Client#workflow_from_hash` and tell `Workflow#initialize`
not to run setup/configure.

Note that for backwards compatibility with workflows persisted before this
change, the setup method will still be called if dependencies in the
deserialized hash are nil.